+++ SoftProcessorUnit/SoftProcessorUnit/SPU/SPU.cpp
@@ -87,7 +87,7 @@ int main(int argc, const char * argv[]) {
         if (params.verbose){
             StackDump(core.stack, -1, params.inputFileName, "ERROR");
             printf("REGs: ");
-            for (int i = 0; i < sizeof(core.REG) / sizeof(core.REG[0]); i++){
+            for (size_t i = 0; i < sizeof(core.REG) / sizeof(core.REG[0]); i++){
                 printf("%lg ", core.REG[i]);
             }
             printf("\n");
